extract sql models
