+++ corehq/apps/app_manager/serializers.py
@@ -3,9 +3,7 @@ from __future__ import unicode_literals
 
 from rest_framework import serializers
 
-from corehq.apps.app_manager.models import (
-    LatestEnabledBuildProfiles,
-)
+from corehq.apps.app_manager.sql_models import LatestEnabledBuildProfiles
 
 
 class LatestEnabledBuildProfileSerializer(serializers.ModelSerializer):
